refactor(tests): consolidate suite and reduce duplicate coverage
Merge small test modules, parametrize overlapping cases, and keep one
core test or UI smoke per scenario so incremental verify stays mapped.
