+++ scripts/verify_targets.py
@@ -34,10 +34,15 @@ CORE_MODULE_TESTS: dict[str, list[str]] = {
     "core/scenario_actions.py": ["tests/test_class_features.py"],
     "core/levels.py": ["tests/test_progression.py"],
     "core/hp_bonuses.py": ["tests/test_progression.py"],
+    "core/dice.py": ["tests/test_stats.py"],
+    "core/constants.py": ["tests/test_stats.py"],
+    "core/races.py": ["tests/test_grants.py"],
+    "core/adventure.py": ["tests/test_models.py"],
+    "core/mod_loader.py": ["tests/test_catalog_loader.py"],
+    "core/skills.py": ["tests/test_proficiencies.py"],
 }
 
 DATA_PATH_TESTS = [
-    "tests/test_mod_loader.py",
     "tests/test_catalog_loader.py",
     "tests/test_io.py",
 ]
@@ -46,17 +51,17 @@ UI_MENU_TESTS: dict[str, str] = {
     "characters_menu": "tests/test_menus_character.py",
     "main_menu": "tests/test_menus_main.py",
     "new_game": "tests/test_menus_character.py",
-    "scenario_flow": "tests/test_scenario.py",
+    "scenario_flow": "tests/test_progression.py",
     "settings": "tests/test_settings.py",
     "backgrounds": "tests/test_backgrounds.py",
     "languages": "tests/test_languages.py",
-    "skills": "tests/test_skills.py",
+    "skills": "tests/test_proficiencies.py",
     "proficiencies": "tests/test_proficiencies.py",
     "expertise": "tests/test_expertise.py",
     "asi": "tests/test_asi.py",
-    "level_up": "tests/test_level_up.py",
+    "level_up": "tests/test_progression.py",
     "class_features": "tests/test_class_features.py",
-    "subclass_trainer": "tests/test_subclass_trainer.py",
+    "subclass_trainer": "tests/test_subclasses.py",
     "_creation_handlers": "tests/test_creation_handlers.py",
     "_deps": "tests/test_menus_character.py",
     "_common": "tests/test_menus_main.py",
