+++ tests/test_menus_character.py
@@ -1,15 +1,28 @@
-"""Тесты UI: выбор персонажа, подрасы, new game, приключения."""
+"""Тесты UI: персонаж, подрасы, new game, creation handlers."""
+
+from typing import Any
 
 import pytest
 
 from core.character_storage import LoadCharactersResult
 from core.models import Adventure, Character
+from core.stats import STAT_NAMES
 from ui.menus import (
     _creation_steps,
     _deps,
     characters_menu,
     new_game,
 )
+from ui.menus._creation_handlers import (
+    _handle_feats,
+    _handle_skills,
+)
+from ui.menus._creation_navigation import (
+    back_step_from_feats,
+    back_step_from_proficiencies,
+    back_step_from_skills,
+)
+from ui.menus._creation_state import _CreationState
 from ui.menus._selectors import select_subrace
 
 
@@ -29,60 +42,78 @@ def _noop_press_enter(monkeypatch: pytest.MonkeyPatch) -> None:
     monkeypatch.setattr(_common, "_press_enter", lambda strings: None)
 
 
-def test_half_orc_auto_selects_single_subrace(
-    monkeypatch, capsys, patch_int_input
-):
-    """Полуорк с одной подрасой выбирается автоматически."""
-    strings = {
-        "character": {
-            "subrace_caption": "ОПИСАНИЕ РАСЫ И ВЫБОР ПОДРАСЫ",
-        }
-    }
-    race = {
-        "name": "Полуорк",
-        "description": "Полуорки — сильные и выносливые воины",
-        "subraces": {
-            "half_orc": {
+def _flat_stats(value: int = 10) -> dict[str, int]:
+    return dict.fromkeys(STAT_NAMES, value)
+
+
+@pytest.mark.parametrize(
+    "race_id,race_data,subrace_strings_fixture,input_values,expected_subrace,"
+    "output_contains",
+    [
+        (
+            "half_orc",
+            {
                 "name": "Полуорк",
                 "description": "Полуорки — сильные и выносливые воины",
-                "ability_bonuses": {"strength": 2},
-            }
-        },
-    }
-
-    monkeypatch.setattr(
-        _deps, "load_race_full", lambda race_id, language="ru": race
-    )
-
-    selected, subrace_id = select_subrace(strings, "half_orc")
-
-    assert selected is True
-    assert subrace_id == "half_orc"
-
-
-def test_human_shows_standard_and_variant_subraces(
+                "subraces": {
+                    "half_orc": {
+                        "name": "Полуорк",
+                        "description": "Полуорки — сильные и выносливые воины",
+                        "ability_bonuses": {"strength": 2},
+                    }
+                },
+            },
+            None,
+            None,
+            "half_orc",
+            None,
+        ),
+        (
+            "human",
+            None,
+            "human_race_with_subraces",
+            [1],
+            "standard",
+            ["Человек (стандарт)", "Человек (вариант)"],
+        ),
+    ],
+)
+def test_select_subrace(
     monkeypatch,
     capsys,
     patch_int_input,
     subrace_strings,
     human_race_with_subraces,
+    race_id,
+    race_data,
+    subrace_strings_fixture,
+    input_values,
+    expected_subrace,
+    output_contains,
 ):
-    """Человек показывает подрасы standard и variant_human."""
-
-    monkeypatch.setattr(
-        _deps,
-        "load_race_full",
-        lambda _race_id, language="ru": human_race_with_subraces,
+    """Выбор подрасы: авто при одной или меню при нескольких."""
+    strings = (
+        subrace_strings
+        if race_id == "human"
+        else {
+            "character": {"subrace_caption": "ОПИСАНИЕ РАСЫ И ВЫБОР ПОДРАСЫ"}
+        }
     )
-    patch_int_input(monkeypatch, [1])
+    race = race_data or human_race_with_subraces
+    monkeypatch.setattr(
+        _deps, "load_race_full", lambda _race_id, language="ru": race
+    )
+    if input_values is not None:
+        patch_int_input(monkeypatch, input_values)
 
-    selected, subrace_id = select_subrace(subrace_strings, "human")
+    selected, subrace_id = select_subrace(strings, race_id)
     output = capsys.readouterr().out
 
     assert selected is True
-    assert subrace_id == "standard"
-    assert "Человек (стандарт)" in output
-    assert "Человек (вариант)" in output
+    assert subrace_id == expected_subrace
+    if output_contains:
+        for fragment in output_contains:
+            assert fragment in output
 
 
 def test_create_character_back_from_subrace_exits(
@@ -117,99 +148,6 @@ def test_create_character_back_from_subrace_exits(
 
     assert result is None
     assert subrace_calls["n"] == 1
-
-
-def test_hardcore_back_from_stats_keeps_rolls(
-    monkeypatch, capsys, ru_strings, patch_int_input
-):
-    """HardCore: «Назад» с подтверждения сохраняет броски 4d6."""
-    from ui.menus.stats import stats_methods
-
-    monkeypatch.setattr(
-        _creation_steps,
-        "select_difficulty",
-        lambda strings: "hardcore",
-    )
-    monkeypatch.setattr(
-        _deps,
-        "get_str_input",
-        lambda *args, **kwargs: "Hero",
-    )
-    monkeypatch.setattr(
-        _deps,
-        "load_races",
-        lambda language="ru": [{"id": "elf", "name": "Эльф"}],
-    )
-    monkeypatch.setattr(
-        _creation_steps,
-        "select_subrace",
-        lambda strings, race_id, language="ru": (True, "wood_elf"),
-    )
-    monkeypatch.setattr(
-        _creation_steps,
-        "select_creation_languages",
-        lambda *args, **kwargs: ["common", "elvish"],
-    )
-
-    def fake_background(*args, **kwargs):
-        return ("folk_hero", ["survival", "animal_handling"])
-
-    monkeypatch.setattr(
-        _creation_steps,
-        "select_creation_background",
-        fake_background,
-    )
-    monkeypatch.setattr(
-        _creation_steps,
-        "select_creation_proficiencies",
-        lambda *args, **kwargs: (
-            ["simple"],
-            ["light"],
-            ["lute", "flute", "drum"],
-        ),
-    )
-    monkeypatch.setattr(
-        _creation_steps,
-        "select_class",
-        lambda strings, language="ru": {"id": "bard"},
-    )
-    monkeypatch.setattr(
-        _creation_steps,
-        "select_creation_skills",
-        lambda *args, **kwargs: ["performance", "persuasion", "deception"],
-    )
-    monkeypatch.setattr(
-        _creation_steps,
-        "select_creation_expertise",
-        lambda *args, **kwargs: ([], []),
-    )
-    monkeypatch.setattr(
-        _creation_steps,
-        "_save_created_character",
-        lambda state: Character(name="Hero", race="elf", class_id="bard"),
-    )
-    monkeypatch.setattr(
-        _creation_steps,
-        "_print_success_and_wait",
-        lambda strings, message: None,
-    )
-    roll_sequence = iter([17, 12, 4, 11, 13, 9, 99, 99, 99, 99, 99, 99])
-    roll_calls = {"count": 0}
-
-    def fake_roll() -> int:
-        roll_calls["count"] += 1
-        return next(roll_sequence)
-
-    monkeypatch.setattr(_deps, "roll_ability_score", fake_roll)
-    monkeypatch.setattr(stats_methods, "_press_enter", lambda strings: None)
-    patch_int_input(monkeypatch, [1, 0, 1])
-
-    _creation_steps.show_create_character_flow(ru_strings)
-    output = capsys.readouterr().out
-
-    assert roll_calls["count"] == 6
-    assert "17" in output
-    assert "99" not in output
 
 
 def test_hardcore_back_to_race_clears_rolls(
@@ -421,9 +359,6 @@ def test_new_game_no_characters_goes_to_create(monkeypatch):
     """Без персонажей — сразу создание."""
     calls = {"select": 0, "create": 0}
 
-    def select_character(strings):
-        calls["select"] += 1
-
     def create_flow(strings, language="ru"):
         calls["create"] += 1
 
@@ -439,38 +374,8 @@ def test_new_game_no_characters_goes_to_create(monkeypatch):
     assert calls == {"select": 0, "create": 1}
 
 
-def test_new_game_back_returns_one_step(monkeypatch):
-    """Назад из выбора приключения возвращает к выбору персонажа."""
-    calls = {"character": 0, "adventure": 0}
-    character = Character(
-        name="Test Hero",
-        race="human",
-        class_id="fighter",
-    )
-
-    def select_character(strings, characters, language="ru"):
-        calls["character"] += 1
-        return character if calls["character"] == 1 else None
-
-    def select_adventure(strings, language, char):
-        calls["adventure"] += 1
-        return None
-
-    monkeypatch.setattr(
-        _deps,
-        "load_characters",
-        lambda: LoadCharactersResult(characters=(character,)),
-    )
-    monkeypatch.setattr(new_game, "_select_character", select_character)
-    monkeypatch.setattr(new_game, "_select_adventure", select_adventure)
-
-    new_game.show_new_game_flow({}, {"language": "ru"})
-
-    assert calls == {"character": 2, "adventure": 1}
-
-
-def test_new_game_back_reuses_cached_character_list(monkeypatch):
-    """Назад из приключения не перечитывает saves без изменений."""
+def test_new_game_back_navigation_and_cached_list(monkeypatch):
+    """Назад из приключения; список персонажей не перечитывается."""
     calls = {"load": 0, "character": 0, "adventure": 0}
     character = Character(
         name="Test Hero",
@@ -576,25 +481,97 @@ def test_characters_menu_delete_all_cancelled(
     assert "Удаление отменено" in output
 
 
-def test_creation_feat_step_routing() -> None:
+@pytest.mark.parametrize(
+    "state_kwargs,expected_feats_required,expected_after_class,expected_back",
+    [
+        (
+            {
+                "name": "Hero",
+                "difficulty": "normal",
+                "race_id": "human",
+                "subrace_id": "variant_human",
+                "class_id": "fighter",
+            },
+            True,
+            "feats",
+            "subclass",
+        ),
+        (
+            {
+                "name": "Hero",
+                "difficulty": "normal",
+                "race_id": "elf",
+                "subrace_id": "wood_elf",
+                "class_id": "fighter",
+            },
+            False,
+            "proficiencies",
+            None,
+        ),
+    ],
+)
+def test_creation_feat_step_routing(
+    state_kwargs: dict[str, Any],
+    expected_feats_required: bool,
+    expected_after_class: str,
+    expected_back: str | None,
+) -> None:
     """Маршрутизация шага черт для variant human vs elf."""
-    variant = _creation_steps._CreationState(
-        name="Hero",
+    state = _creation_steps._CreationState(**state_kwargs)
+    assert (
+        _creation_steps._feats_step_required(state) is expected_feats_required
+    )
+    assert (
+        _creation_steps._step_after_class_choice(state) == expected_after_class
+    )
+    if expected_back is not None:
+        assert _creation_steps._back_step_from_feats(state) == expected_back
+
+
+def test_back_step_from_skills_and_feats_without_class() -> None:
+    state = _CreationState(name="Test", difficulty="normal")
+    assert back_step_from_skills(state) == "proficiencies"
+    assert back_step_from_feats(state) == "class"
+
+
+def test_back_step_from_feats_with_subclass(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    state = _CreationState(
+        name="Test",
+        difficulty="normal",
+        class_id="fighter",
+    )
+    monkeypatch.setattr(
+        "ui.menus._creation_navigation.subclass_offered_at_creation",
+        lambda _difficulty, _class_id: True,
+    )
+    assert back_step_from_feats(state) == "subclass"
+
+
+def test_handlers_without_class_id() -> None:
+    state = _CreationState(
+        name="Test",
+        difficulty="normal",
+        race_id="human",
+        stats=_flat_stats(),
+    )
+    assert _handle_feats({}, state, "ru").next_step == "class"
+    assert _handle_skills({}, state, "ru").next_step == "proficiencies"
+
+
+def test_back_step_from_proficiencies_via_feats(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    state = _CreationState(
+        name="Test",
         difficulty="normal",
         race_id="human",
         subrace_id="variant_human",
         class_id="fighter",
     )
-    assert _creation_steps._feats_step_required(variant)
-    assert _creation_steps._step_after_class_choice(variant) == "feats"
-    assert _creation_steps._back_step_from_feats(variant) == "subclass"
-
-    elf = _creation_steps._CreationState(
-        name="Hero",
-        difficulty="normal",
-        race_id="elf",
-        subrace_id="wood_elf",
-        class_id="fighter",
+    monkeypatch.setattr(
+        "ui.menus._creation_navigation.feats_step_required",
+        lambda _state: True,
     )
-    assert not _creation_steps._feats_step_required(elf)
-    assert _creation_steps._step_after_class_choice(elf) == "proficiencies"
+    assert back_step_from_proficiencies(state) == "feats"
